% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c40_colors.R
\name{c40_colors}
\alias{c40_colors}
\title{Function to get C40 colors}
\usage{
c40_colors(..., name = FALSE)
}
\arguments{
\item{...}{Name or number of the color. The available c40 colors can be viewed with `avalilable_colors()`.}

\item{name}{If TRUE returns the name of the hex color}
}
\value{
C40 color pallets
}
\description{
Function to get C40 colors
}
\examples{
### Get c40 yellow color
c40_colors("yellow", name = TRUE)

c40_colors("yellow", name = FALSE)
}
