% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c40_pallets.R
\name{c40_palettes}
\alias{c40_palettes}
\title{Get C40 Colour Palettes}
\usage{
c40_palettes(palette = "qualitative", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character. Name of the colour palette. Options are:
\describe{
  \item{"qualitative"}{For categorical data (7 distinct colours)}
  \item{"sequential"}{For ordered data (light to dark blue)}
  \item{"dicotomic"}{For binary data (red/green) - note: kept for
    backwards compatibility, preferred spelling is "dichotomous"}
  \item{"divergent"}{For data with meaningful midpoint (red-white-green)}
}}

\item{reverse}{Logical. If \code{TRUE}, reverses the palette order.
Default is \code{FALSE}.}

\item{...}{Additional arguments passed to \code{grDevices::colorRampPalette()}.}
}
\value{
A function that takes an integer \code{n} and returns \code{n} colours.
}
\description{
Returns a colour palette function based on C40 brand guidelines.
The returned function can be used to generate any number of colours
by interpolation.
}
\examples{
# Get qualitative palette function
pal <- c40_palettes("qualitative")
pal(5)  # Generate 5 colours

# Use with ggplot2
\dontrun{
library(ggplot2)
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
  geom_point() +
  scale_colour_manual(values = c40_palettes("qualitative")(3))
}

}
\seealso{
\code{\link{scale_fill_c40}}, \code{\link{scale_color_c40}},
  \code{\link{c40_colours}}
}
