% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c40_pallets.R
\name{c40_pallets}
\alias{c40_pallets}
\title{Get the C40 color pallets}
\usage{
c40_pallets(palette = "qualitative", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of the color pallet. The available color pallets can be viewed with `available_pallets()`}

\item{reverse}{if *TRUE* returns inverted color pallet}

\item{...}{Aditional arguments to define with colorRampPalette()}
}
\value{
C40 official color pallets
}
\description{
Get the C40 color pallets
}
