% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ghg_emissions.R
\name{calc_ghg_emissions}
\alias{calc_ghg_emissions}
\title{Calculate Greenhouse Gas (GHG) Emissions by Region, Tier, or City}
\usage{
calc_ghg_emissions(region = "all", sum_by = "city")
}
\arguments{
\item{region}{A character string specifying the region to filter by. Valid options include:
\code{"all"}, \code{"Africa"}, \code{"Central East Asia"}, \code{"East, Southeast Asia and Oceania"},
\code{"Europe"}, \code{"Latin America"}, \code{"North America"}, and \code{"South and West Asia"}.
The default is \code{"all"}, which includes all regions.}

\item{sum_by}{A character string specifying the grouping for the summary. Options include:
\code{"region"}, \code{"tier"}, or \code{"city"}. \code{"region"} groups results by year and region,
\code{"tier"} groups by year, tier, and tier description, and \code{"city"} groups by year, region, and city.}
}
\value{
A data frame containing the following columns:
\itemize{
  \item \code{year}: The year of the emissions data.
  \item \code{region}: The region of the emissions data (if applicable).
  \item \code{city}: The city (if \code{sum_by = "city"} is used).
  \item \code{population}: The total population for the grouping.
  \item \code{target_emissions_tco2e}: The target emissions in tonnes of CO2 equivalent.
  \item \code{actual_emissions_tco2e}: The actual emissions in tonnes of CO2 equivalent.
  \item \code{percapita_actual}: Actual emissions per capita.
  \item \code{percapita_target}: Target emissions per capita.
  \item \code{gap_percapita}: The percentage gap between actual and target per capita emissions.
  \item \code{gap_absolute}: The percentage gap between actual and target absolute emissions.
  \item \code{gap_percapita_color}: A color-coded classification of the \code{gap_percapita}.
}
}
\description{
This function calculates and summarizes greenhouse gas (GHG) emissions for specified regions,
capacity tiers, or cities based on modeled and target emissions data from the C40 Data Warehouse.
It retrieves data from the database, performs necessary data processing, and outputs emissions
per capita and other related metrics based on the chosen grouping.
}
\details{
The function establishes a connection to the C40 Data Warehouse and retrieves
emissions data from multiple database tables. The data is then filtered by the specified
\code{region} and joined with existing tier data in the package. The grouping and summarization
depend on the \code{sum_by} parameter, and the final output includes calculated metrics to
help evaluate emission targets.
}
\examples{
\dontrun{
# Calculate emissions for all regions, grouped by region
calc_ghg_emissions(region = "all", sum_by = "region")

# Calculate emissions for Europe, grouped by tier
calc_ghg_emissions(region = "Europe", sum_by = "tier")

# Calculate emissions for North America, grouped by city
calc_ghg_emissions(region = "North America", sum_by = "city")
}

}
