% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_city_names.R
\name{check_and_correct_cities}
\alias{check_and_correct_cities}
\title{Clean and correct city names}
\usage{
check_and_correct_cities(
  df,
  city_column_name,
  valid_cities,
  want_cityid = TRUE,
  threshold = 0.2
)
}
\arguments{
\item{df}{dataframe with the city column name}

\item{city_column_name}{name of the column wich has the city names}

\item{valid_cities}{C40 official City names}

\item{want_cityid}{TRUE/FALSE if you want to return a dataframe with a city_id column, in order to link it with the Data Warehouse. Default as FALSE}

\item{threshold}{matching level}
}
\value{
If clean_city_name = FALSE, returns just comparisson between the current city column names and the C40 official city names in the Data Warehouse.
If clean_city_name = TRUE, returns a dataframe with the correct and official city names
}
\description{
Clean and correct city names
}
