% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Clean text function:}
\usage{
clean_text(text)
}
\arguments{
\item{text}{character value}
}
\description{
Removes accents and umlaut from vowels. E.g.: "ÿÚòâ" into "yUoa".
}
\examples{
clean_text("ÿÚòâ") #"yuoa"

}
