% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_tiers}
\alias{df_tiers}
\title{Data Frame for Capacity Tiers}
\format{
A data frame with X rows and Y columns.
}
\usage{
df_tiers
}
\description{
A data frame that contains information on capacity tiers.
}
\keyword{datasets}
