% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_c40cities.R
\name{get_c40cities}
\alias{get_c40cities}
\title{Get current and past C40 cities}
\usage{
get_c40cities(current = TRUE)
}
\value{
A dataset with the City, City id and if it is currently part of C40
}
\description{
Get current and past C40 cities
}
\examples{
df_c40_cities <- get_c40cities(current = TRUE)
}
