% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dw_db_example.R
\name{get_dw_db_example}
\alias{get_dw_db_example}
\title{Code examples for accesing the data from the Data Warehouse}
\usage{
get_dw_db_example(type = "table")
}
\arguments{
\item{type}{Type of request (sql script, table)}
}
\value{
a text with the code to see how the Data Warehouse can be accessed
}
\description{
Code examples for accesing the data from the Data Warehouse
}
\examples{
get_dw_db_example(type = 'sql')
}
