% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_googledrive_path.R
\name{get_googledrive_path}
\alias{get_googledrive_path}
\title{Get your local Google Drive folder}
\usage{
get_googledrive_path()
}
\value{
A path to your local google drive folder
}
\description{
Get your local Google Drive folder
}
\examples{
path <- get_googledrive_path()
}
