% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_c40.R
\name{scale_fill_c40}
\alias{scale_fill_c40}
\title{C40 Scale color for filling}
\usage{
scale_fill_c40(palette = "qualitative", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{pallet name from `c40_pallets()`. Options can be viewed in `available_pallet()`}

\item{discrete}{*TRUE* for discrete variables}

\item{reverse}{*TRUE* returns inverted colors}

\item{...}{Aditional arguments for discrete_scale() or
scale_color_gradientn()}
}
\description{
C40 Scale color for filling
}
