% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ghg.R
\name{calc_ghg}
\alias{calc_ghg}
\title{Obtain GHG inventory and modelled emissions}
\usage{
calc_ghg(city = "all", year, source = "both")
}
\value{
A dataset with the emissions figures
}
\description{
Obtain GHG inventory and modelled emissions
}
\examples{
df_ghg_emissions <- get_ghg_emissions(city = c("Oslo", "Copenhagen", "Stockholm"),
                                      year = 2000:2023,
                                      source = "both")
}
