% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_city_names.R
\name{check_city_names}
\alias{check_city_names}
\title{Check if city column names matches with the C40 curated list of city names}
\usage{
check_city_names(
  df,
  city_var_name,
  clean_city_name = FALSE,
  want_cityid = FALSE
)
}
\arguments{
\item{df}{dataframe with the city column name}

\item{city_var_name}{name of the column wich has the city names}

\item{clean_city_name}{TRUE/FALSE if you want to run a cleaning process for city names. Default as FALSE}

\item{want_cityid}{TRUE/FALSE if you want to return a dataframe with a city_id column, in order to link it with the Data Warehouse. Default as FALSE}
}
\value{
If clean_city_name = FALSE, returns just comparisson between the current city column names and the C40 official city names in the Data Warehouse.
If clean_city_name = TRUE, returns a dataframe with the correct and official city names
}
\description{
Check if city column names matches with the C40 curated list of city names
}
